drop fixes/nullptr-dereference.patch, as it's only needed for older compilers
The patch is for a codepath that's never reached after clang 9. It can safely
be deleted, since sid has clang 13, bullseye has clang 11, and buster has
clang 11 backported to it.

The path is an #else branch of an #ifdef SUPPORTS_LOCATION_BUILTINS, and that
macro is defined if either gcc >= 7 or clang supports the
__builtin_{FUNCTION,LINE,FILE} builtins (which was added to the docs
for llvm 9).