bundle some nodejs modules to allow optimize_webui.py to run during build
Even with optimize_webui=false in debian/rules, optimize_webui.py is
needed for the build. Specifically, this commit made it necessary:

https://chromium.googlesource.com/chromium/src/+/742509d9c81c2cdf89d786b2a5374f727e032e34%5E%21/

That python script depends on some nodejs modules, some of which
aren't in debian. For the ones that are, add build-deps. For the
ones that aren't, stop deleting them from
third_party/node/node_modules when generating the orig tarball.

At some point it might be easier to just use all the bundled
node_modules, but for right now we'll use debian's where available.