drop system/use-desktop-gl-as-default.patch
The compositor's buffer would fill up, resulting in the tab crashing with this
error:

[231882:8:0101/114936.340325:FATAL:presentation_time_callback_buffer.cc(110)] Check faile
d: frame_token_infos_.size() <= kMaxBufferSize (61 vs. 60)

Doubling kMaxBufferSize didn't help; but it turned out that overriding the
desktop GL (via --use-gl=any) did work. I'm not sure if the slow draw issues
are related to enabling ozone or just changes in the newer version, but
either way I'm not convinced that changing the default is a good idea.