enable official builds
This will get rid all the crashes caused by DCHECK (debug checks) that
are fatal errors when they should really be debugging output. It's also
what all the other distributions do (Arch, Fedora, Gentoo, etc).

Note that for now we're disabling swiftshader, but we might want to
enable and ship it.