drop disable/buildbot.patch, disable/mojo.patch, & disable/web-ui-tests.patch
disable/buildbot.patch was needed because we delete the buildbot directory
in the chromium source. Debian doesn't actually use upstream's buildbot
stuff, so at some point mgilbert decided to delete the whole directory.
However, it's not worth deleting it when we also need to maintain a patch
to fix the resulting build issues.

disable/mojo.patch appears to have been for past build issues due to files
missing from the tarball (but it exists in git). We're adding a new patch
that is extremely large, but it's literally just a copy of all the
missing files from upstream git and is hopefully only temporary. Upstream
bug report https://bugs.chromium.org/p/chromium/issues/detail?id=1293630
will determine what we do around this.

disable/web-ui-tests.patch is a supporting patch for the mojo stuff. With
the files re-added, this is also no longer necessary.