drop most of bullseye/icu-types.patch
A bunch of this code wasn't actually needed for supporting icu-67 (stable).
Instead, it was a patch taken from upstream that allows switching underlying
uchar16 types:

https://chromium.googlesource.com/chromium/src/+/d9ef4db129efab6b8c7da5d81ba608124c65d052%5E%21/
https://bugs.chromium.org/p/chromium/issues/detail?id=911896#c71

That type migration finished up, and the above commit was basically reverted:
https://chromium.googlesource.com/chromium/src/+/43d5378f7f363dab9271ca37774c71176c9e7b69%5E!/

So, we can safely drop that part of this patch.

The rest of the patch is of course needed for API changes between icu-67
and icu-69.