add manual dependencies on xdg-desktop-portal-* packages
Upstream used to link against libgtk-3-0 at compile time. However,
in order for them to ship a single binary that supports both gtk3 and
gtk4, they stopped doing that. Instead, at runtime chromium detects
whether gtk3 or gtk4 is available and manually dlopen()s the
appropriate .so library file.

This breaks debian's build tools, which detects what a binary
is compiled against and automatically includes the necessary library
dependencies. So we need to add a manual dependency on libgtk; and
specifically, we'll depend upon xdg-desktop-portal packages so that
KDE users don't need to install libgtk with chromium.