drop disable/glmark.patch
This patch is unnecessary. It was originally included because we
delete the entire third_party/angle/third_party/glmark2/ directory.
However, all the dependencies that references that directory are
inside of an "if (build_angle_perftests) {...}", and
build_angle_perftests defaults to false. No other configure
options touch it, so we're safe just ensuring that it's manually
set to false and dropping this patch.