drop disable/fuzzers.patch and adjust debian/copyright
With the last release, we stopped deleting all the fuzz directories. That
means disable/fuzzers.patch is no longer necessary, so we can drop it.

Haha, I lied! Actually, there's one last thing that need to change in
order to drop it - third_party/zlib/contrib/ contains a fuzz directory,
and we were manually deleting it. The contrib directory contains a
mix of source code, some of it from chromium (and under chromium's
BSD-3-clause license), and the rest of it from zlib. The zlib code is
under zlibs license, so it's all good. So we also adjust debian/copyright
to stop dropping that directory.