release preparing 102.0.5005.115-1, w/ fixes
Needed two separate sets of 32-bit build fixes here.

1) On bullseye, builds were failing on armhf and i386 due to a an
issue w/ constexpr in sys_byteswap.h. I fixed that in the prior
release (and they fixed it upstream), but it turns out that there's
a second copy of sys_byteswap.h that also is broken - so I needed to
fix that on also. #1011096

2) Upstream disabled support for i386 builds on Linux. They didn't
want to add an override (*sigh*), so we need to patch out the assert()
that kills the build if linux && x86 is detected.