start preparing 105.0.5195.52
Mostly normal refreshes and build fixes, but: libevent changed
location from base/third_party/ to third_party/, and a bunch of
new libraries can be linked to externally.

For now, we're continuing to use the bundled libraries. I'll have
to test them each to see which ones are already in debian, and work
in both sid and bullseye.