rework default search engine stuff
Instead of using policy (which displays scary "your browser is being
managed by an organization" messages in the browser), use
master_preferences to default to DuckDuckGo for the search engine.

That message will disappear. However, the downside is that the
default search engine will only be set on the initial installation
of chromium.

The duckduckgo.json file is treated as a conffile, so if the user
edited it it will remain.