drop disable/installer.patch
We delete chome/install_static, and d/patches/disable/installer.patch
is needed in order to work around the missing BUILD.gn and a header
file. There's no real need to delete install_static, as it follows
the same license as the rest of chromium, it's tiny (260 kB), and
nothing in there other than the buildflags.h header is actually built
on linux. The entire thing is protected with is_win/IS_WIN checks.

Chromium 106 also moved the //chrome/installer and //chrome/updater
bits elsewhere, and that doesn't appear to be pulled in either. The
only stuff in there that gets built is stuff in
chrome/installer/util/, which is specified elsewhere (in
chrome/BUILD.gn).

tl;dr: there's no real need for this patch, the installer doesn't
get build either way.