start building libaom on armhf again
Previously we got this without disable/libaom-arm.patch:
https://buildd.debian.org/status/fetch.php?pkg=chromium&arch=armhf&ver=99.0.4818.0-0.1&stamp=1642756579&raw=0

arm_cpudetect.c (which contains the definition of aom_arm_cpu_caps())
was included for arm64 and arm32 neon builds, but not for arm32
vfpv3-d16 builds. It just needed to be included for ALL arm builds
regardless of version, so that's what I'm going to try here (and
then send upstream if it works).