disable qt5, and update a bunch of patches for v107
Upstream now builds against QT5; see
https://groups.google.com/a/chromium.org/g/chromium-packagers/c/-2VGexQAK6w
We're going to disable this for now, and wait for the KDE folks to chime
in on whether this is desirable enough to add a large chunk of migration deps.

Patch updates are mostly basic refreshes or updates, but see the
clang-version-check.patch header for more details about that issue.

  - disable/catapult.patch: delete add'l blink reference to catapult.
  - bullseye/clang13.patch: refresh for minor upstream changes.
  - ppc64le/workarounds/HACK-third_party-libvpx-use-generic-gnu.patch: refresh
  - disable/clang-version-check.patch: added to fix build failure. Needs
    to go upstream.
  - ppc64le/workarounds/HACK-debian-clang-disable-skia-musttail.patch:
    drop, upstream skia stopped using clang::musttail.