scripts/get-cves.sh: automatically wrap lines at 78 chars
Rather than breaking the line up arbitrarily, we stick the "Reported by"
piece on the next line. This doesn't always mean that the two split lines
will fit into 80 chars, but it's good enough for most cases and those corner
cases need to be manually edited either way.

And hey, might as well update the changelog for the automatically-wrapped
cve output.