fix disable/unrar.patch and take ownership of it
This is a completely rewritten patch, so adjust the authorship.

LOG(ERROR) instead of FATAL so we don't die and leave stuff unfinished.

Also, rar downloads would previously be left in an incomplete
state, so you couldn't even download them. Now we complete the download
(without the safety check) so that you can at least unarchive the rar
with your preferred (free?) software.