revert upstream change using new brotli features
I don't know if we should file a wishlist bug against debian's brotli
asking for a git snapshot, or if we should just be patient and wait
for brotli to do a new release. If chromium keeps making changes to
this code, we might have to switch to bundling brotli.