Add a bunch more patches to get bad-font-gc2.patch working
Along with just getting the patches to apply, there's a bunch of
additional stuff that was changed upstream from using refptrs to handling
deleting through garbage collection that shared code w/
FontPlatformData. Working around it resulted in crashes, so I
thought it best to just include all of these commits for now.