preserve chromium arguments
We had a problem with argument processing. Options that we
specifically wanted to process in the script before starting chromium
were ignored if chromium arguments came first. This fixes that, by
skipping over (but preserving) chromium arguments to process ALL
arguments. Once we get to non-arguments (eg, a URL), we're done scanning
and can move on.