Fix handling of quoted arguments
Unknown args can include quotes, so by saving to $CHROMIUM_FLAGS and
passing around by strings we break things for some users. Instead, keep
those unknown args on the stack (pushing to the back), and just track
how many we have. When all we have left remaining are unknown args, it's
safe to stop processing args.

Later on, we run the chromium binary with "$@", which ensures all the
quotes don't disappear.