drop system/clang-format.patch
When we switched to using clang-16 (instead of the default "clang"),
this patch (which calls "clang-format" instead of "clang-format-16")
should've broken. It didn't. I'm assuming mgilbert originally added
it due to a build error, but he didn't document it so I'm just guessing;
either way, it would appear that we don't need it.