more 128 build fixes
Update clang16.patch to remove another unsupported clang build option.

highway-include-path.patch was a simple header fix, but unfortunately
not only is upstream arguing against fixing it (in
https://issues.chromium.org/issues/357975577 ), they've also had someone
else "fix" it indepedently in a more complicated manner:
https://chromium-review.googlesource.com/c/chromium/src/+/5667363
which, tbh, I haven't looked closely at, but it seems like a rather
bizarre change.

So we need to undo that change as well in highway-include-path.patch.