Set use_wayland_gbm=false
Explanation from JC:
The issue is that the code assumes that if you have wayland then the
gbm allocator associated with it should be able to allocate dmabufs of
all the sorts that wayland can consume. Unfortunately our GBM can
still only allocate RGB-like single planar buffers and so when Chrome
wants a YUV buffer it all fails. The above switch tells the code this
assumption is wrong.

It would be possible, with only moderate effort, to write an allocator
that uses udmabuf or dma_heap.

The ideal solution is that our Vulkan makes it to a point that Chrome
will accept, then we can use Dawn, Graphite etc. and Chrome will use
Vulkans allocator which really should deal with YUV, as well as being
generally more efficient.